# src/entities/User.ts
import { IsEmail, IsNotEmpty } from "class-validator";

import type { ObjectID } from "@/libs/typeorm";
import { Column, Entity, ObjectIdColumn } from "@/libs/typeorm";

import Base from "./Base";
import type { Role } from "./Role";
import type { Team } from "./Team";
import type { Workspace } from "./Workspace";

export interface ProviderInfo {
	name: string;
	user_id?: string;
	access_token?: string;
}

export interface AccessTokenInfo {
	access_token: string;
	expiredTimestamp: number;
	expiredDate: Date;
	expiredDateGTM7: string;
}

@Entity({ name: "users" })
export default class User extends Base {
	/**
	 * User name
	 */
	@Column({ length: 250 })
	@IsNotEmpty({ message: `User name is required.` })
	name?: string;

	/**
	 * Unique username of a user
	 * This equavilent with "slug"
	 */
	@Column()
	username?: string;

	/**
	 * Service Account is User!
	 */
	@Column({ default: "user" })
	type?: "user" | "service_account";

	/**
	 * User email address
	 */
	@Column({ length: 500 })
	@IsNotEmpty({ message: `Email is required.` })
	@IsEmail(null, { message: `Email is not valid.` })
	email?: string;

	/**
	 * Is this user's email or phone verified?
	 */
	@Column("boolean", { default: false })
	verified?: boolean;

	/**
	 * User profile picture URL
	 */
	@Column({ type: "string" })
	image?: string;

	/**
	 * List of Cloud Providers which this user can access to
	 */
	@Column({ array: true })
	providers?: ProviderInfo[];

	/**
	 * User password (hashed)
	 */
	@Column({ type: "string" })
	password?: string;

	/**
	 * User token
	 */
	@Column()
	token?: AccessTokenInfo;

	/**
	 * User's roles
	 * TODO: Thêm "workspace" cho "roles" vì mỗi workspace sẽ có roles khác nhau, trong khi mỗi user có thể nằm trong nhiều workspace
	 */
	@ObjectIdColumn({ name: "roles", array: true, default: [] })
	roles?: (ObjectID | Role | string)[];

	/**
	 * User's team IDs which this user is a member
	 */
	@ObjectIdColumn({ name: "teams", array: true, default: [] })
	teams?: (ObjectID | Team | string)[];

	/**
	 * List of workspace IDs which this user is a member
	 */
	@ObjectIdColumn({ name: "workspaces", array: true, default: [] })
	workspaces?: (ObjectID | Workspace | string)[];

	/**
	 * Active workspace of a user
	 */
	@ObjectIdColumn({ name: "workspaces" })
	activeWorkspace?: ObjectID | Workspace | string;

	constructor(data?: User) {
		super();
		Object.assign(this, data);
	}
}

export { User };